@@ -4,55 +4,55 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.1.3",
-    "@angular/common": "13.1.3",
-    "@angular/compiler": "13.1.3",
-    "@angular/core": "13.1.3",
-    "@angular/forms": "13.1.3",
-    "@angular/platform-browser": "13.1.3",
-    "@angular/platform-browser-dynamic": "13.1.3",
-    "@angular/router": "13.1.3",
-    "@nrwl/angular": "13.4.1",
+    "@angular/animations": "14.1.0",
+    "@angular/common": "14.1.0",
+    "@angular/compiler": "14.1.0",
+    "@angular/core": "14.1.0",
+    "@angular/forms": "14.1.0",
+    "@angular/platform-browser": "14.1.0",
+    "@angular/platform-browser-dynamic": "14.1.0",
+    "@angular/router": "14.1.0",
+    "@nrwl/angular": "14.4.3",
     "date-fns": "2.29.1",
     "rxjs": "7.4.0",
     "tslib": "2.3.1",
     "zone.js": "0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.1.4",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.1.4",
-    "@angular/compiler-cli": "13.1.3",
-    "@angular/language-service": "13.1.3",
-    "@nrwl/cli": "13.10.6",
-    "@nrwl/cypress": "13.10.6",
-    "@nrwl/eslint-plugin-nx": "13.10.6",
-    "@nrwl/jest": "13.10.6",
-    "@nrwl/linter": "13.10.6",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/tao": "13.10.6",
-    "@nrwl/workspace": "13.10.6",
+    "@angular-devkit/build-angular": "14.1.0",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.1.0",
+    "@angular/language-service": "14.1.0",
+    "@nrwl/cli": "14.4.3",
+    "@nrwl/cypress": "14.4.3",
+    "@nrwl/eslint-plugin-nx": "14.4.3",
+    "@nrwl/jest": "14.4.3",
+    "@nrwl/linter": "14.4.3",
+    "@nrwl/nx-cloud": "14.2.0",
+    "@nrwl/tao": "14.4.3",
+    "@nrwl/workspace": "14.4.3",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.45",
-    "@typescript-eslint/eslint-plugin": "5.8.1",
+    "@typescript-eslint/eslint-plugin": "5.31.0",
     "cypress": "9.2.1",
-    "eslint": "8.5.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-cypress": "2.12.1",
-    "jest": "27.4.7",
-    "jest-preset-angular": "11.0.1",
-    "prettier": "2.5.1",
+    "jest": "28.1.3",
+    "jest-preset-angular": "12.2.0",
+    "prettier": "2.7.1",
     "tailwindcss": "3.1.6",
-    "ts-jest": "27.1.5",
-    "typescript": "4.4.4"
+    "ts-jest": "28.0.7",
+    "typescript": "4.7.4"
   }
 }
